Update imputed_merged_vcf.py
Change script to utilize _imputed_parents.txt files instead of hVCF files. This was done because hVCF files may have haplotypes in the headers where the reference range assigned to the parent does not match how it is used in the data lines of the file. Additionally, there can be discrepencies between the _imputed_parents.txt files and hVCFs. This can occur if a haplotype ID is used more than once between different parents. While the sequence may be the same, the loci in the VCF are not. For these reasons, I decided to use _imputed_parents.txt files instead, which have the imputed parent at each reference range for imputed lines.